requirements.txt
dnspython
requests
BeautifulSoup4
censys